More cleanup, added text icons

+++ src/main/java/dev/latvian/mods/kubejs/item/custom/SwordItemBuilder.java
@@ -7,7 +7,7 @@ import net.minecraft.world.item.SwordItem;
 public class SwordItemBuilder extends HandheldItemBuilder {
 	public SwordItemBuilder(ResourceLocation i) {
 		super(i, 3F, -2.4F);
-		itemAttributeModifiers = SwordItem.createAttributes(toolTier, (int) attackDamageBaseline, speedBaseline);
+		itemAttributeModifiers = SwordItem.createAttributes(toolTier, attackDamageBaseline, speedBaseline);
 		parentModel = "minecraft:item/iron_sword";
 	}
